# mobile/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "description": "Create a new user account with email and password",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "User registration data",
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterDto" },
              "examples": {
                "example1": {
                  "summary": "Example registration",
                  "description": "A sample user registration",
                  "value": {
                    "first_name": "John Doe",
                    "last_name": "Doe",
                    "email": "john.doe@example.com",
                    "password": "securePassword123",
                    "confirm_password": "securePassword123",
                    "role": "Pupil"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponseDto" }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponseDto" },
                "example": {
                  "statusCode": 400,
                  "message": [
                    "email must be a valid email",
                    "password must be longer than or equal to 6 characters"
                  ],
                  "error": "Bad Request"
                }
              }
            }
          },
          "409": {
            "description": "User already exists",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponseDto" },
                "example": {
                  "statusCode": 409,
                  "message": "User already exists",
                  "error": "Conflict"
                }
              }
            }
          }
        },
        "summary": "Register a new user",
        "tags": ["Authentication"]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Authenticate user with email and password",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "User login credentials",
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDto" },
              "examples": {
                "example1": {
                  "summary": "Example login",
                  "description": "A sample user login",
                  "value": {
                    "email": "john.doe@example.com",
                    "password": "securePassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged in",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponseDto" },
                "example": {
                  "message": "User successfully logged in",
                  "data": {
                    "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "refresh_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "user": {
                      "id": "123e4567-e89b-12d3-a456-426614174000",
                      "email": "john.doe@example.com",
                      "name": "John Doe"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponseDto" },
                "example": {
                  "statusCode": 400,
                  "message": [
                    "email must be a valid email",
                    "password should not be empty"
                  ],
                  "error": "Bad Request"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponseDto" },
                "example": {
                  "statusCode": 401,
                  "message": "Invalid credentials",
                  "error": "Unauthorized"
                }
              }
            }
          }
        },
        "summary": "User login",
        "tags": ["Authentication"]
      }
    },
    "/auth/google/token": {
      "post": {
        "operationId": "AuthController_exchangeGoogleIdToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleExchangeTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponseDto" }
              }
            }
          }
        },
        "summary": "Exchange Google id_token for app tokens",
        "tags": ["Authentication"]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Get a new access token using refresh token",
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Refresh token",
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshTokenDto" },
              "examples": {
                "example1": {
                  "summary": "Example refresh",
                  "description": "A sample refresh token request",
                  "value": {
                    "refresh_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponseDto" }
              }
            }
          },
          "401": {
            "description": "Invalid refresh token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponseDto" },
                "example": {
                  "statusCode": 401,
                  "message": "Invalid refresh token",
                  "error": "Unauthorized"
                }
              }
            }
          }
        },
        "summary": "Refresh access token",
        "tags": ["Authentication"]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "description": "Send a password reset link to user email",
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "User email for password reset",
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ForgotPasswordDto" },
              "examples": {
                "example1": {
                  "summary": "Example forgot password",
                  "description": "A sample forgot password request",
                  "value": { "email": "john.doe@example.com" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset email sent (if email exists)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponseDto" }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponseDto" },
                "example": {
                  "statusCode": 400,
                  "message": ["email must be a valid email"],
                  "error": "Bad Request"
                }
              }
            }
          }
        },
        "summary": "Request password reset",
        "tags": ["Authentication"]
      }
    },
    "/auth/reset-password": {
      "post": {
        "description": "Reset user password using reset token",
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Reset token and new password",
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResetPasswordDto" },
              "examples": {
                "example1": {
                  "summary": "Example reset password",
                  "description": "A sample password reset",
                  "value": {
                    "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "new_password": "newSecurePassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponseDto" },
                "example": { "message": "Password reset successfully" }
              }
            }
          },
          "400": {
            "description": "Invalid or expired reset token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponseDto" },
                "example": {
                  "statusCode": 400,
                  "message": "Invalid or expired reset token",
                  "error": "Bad Request"
                }
              }
            }
          }
        },
        "summary": "Reset password",
        "tags": ["Authentication"]
      }
    },
    "/auth/request-email-verification": {
      "post": {
        "description": "Request email verification token",
        "operationId": "AuthController_requestEmailVerification",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Email verication sent.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponseDto" },
                "example": {
                  "message": "If the email exists, an email verification link has been sent."
                }
              }
            }
          }
        },
        "security": [{ "JWT-auth": [] }],
        "summary": "Request email verification",
        "tags": ["Authentication"]
      }
    },
    "/auth/verify-email": {
      "get": {
        "description": "Verify user email using verification token",
        "operationId": "AuthController_verifyEmail",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "description": "Email verification token",
            "schema": {
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email verified successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponseDto" },
                "example": { "message": "Email verified successfully" }
              }
            }
          },
          "400": {
            "description": "Invalid or expired verification token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponseDto" },
                "example": {
                  "statusCode": 400,
                  "message": "Invalid or expired verification token",
                  "error": "Bad Request"
                }
              }
            }
          }
        },
        "summary": "Verify email",
        "tags": ["Authentication"]
      }
    },
...
    "/auth/change-password": {
      "post": {
        "description": "Change the authenticated user password",
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Current and new password",
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangePasswordDto" },
              "examples": {
                "example1": {
                  "summary": "Example change password",
                  "description": "A sample password change",
                  "value": {
                    "current_password": "currentPassword123",
                    "new_password": "newSecurePassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponseDto" },
                "example": { "message": "Password changed successfully" }
              }
            }
          },
          "401": {
            "description": "Invalid or missing token / Current password incorrect",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponseDto" },
                "example": {
                  "statusCode": 401,
                  "message": "Current password is incorrect",
                  "error": "Unauthorized"
                }
              }
            }
          }
        },
        "security": [{ "JWT-auth": [] }],
        "summary": "Change user password",
        "tags": ["Authentication"]
      }
    },
...
    "/auth/verify-token": {
      "get": {
        "description": "Verify if the provided JWT token is valid and return user info",
        "operationId": "AuthController_verifyToken",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Token is valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": { "type": "boolean", "example": true },
                    "user": { "$ref": "#/components/schemas/UserResponseDto" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponseDto" },
                "example": {
                  "statusCode": 401,
                  "message": "Unauthorized",
                  "error": "Unauthorized"
                }
              }
            }
          }
        },
        "security": [{ "JWT-auth": [] }],
        "summary": "Verify JWT token",
        "tags": ["Authentication"]
      }
    },
...
  },
  "info": {
    "title": "LexiLearner API",
    "description": "",
    "version": "1.0",
    "contact": {},
    "license": { "name": "MIT", "url": "https://opensource.org/licenses/MIT" }
  },
...
  "servers": [
    { "url": "http://localhost:3000", "description": "Development server" }
  ],
  "components": {
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Username",
            "example": "johndoe1",
            "minLength": 8
          },
          "first_name": {
            "type": "string",
            "description": "User first name",
            "example": "John",
            "minLength": 1
          },
          "last_name": {
            "type": "string",
            "description": "User last name",
            "example": "Doe",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "description": "User email",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "securePassword123"
          },
          "confirm_password": {
            "type": "string",
            "description": "Password confirmation",
            "example": "securePassword123"
          },
          "role": {
            "type": "string",
            "description": "User role",
            "example": "Pupil",
            "enum": ["Pupil", "Teacher"]
          }
        },
        "required": [
          "username",
          "first_name",
          "last_name",
          "email",
          "password",
          "confirm_password",
          "role"
        ]
      },
...
    }
  }
}